+++ main/modules/network/esp3d_network.cpp
@@ -684,7 +684,9 @@ bool Esp3DNetwork::stopBtMode() {
 }
 
 bool Esp3DNetwork::setMode(Esp3dRadioMode mode, bool restart) {
-  esp3d_log("Current mode is %d, and ask for %d", _current_radio_mode, mode);
+  esp3d_log("Current mode is %d, and ask for %d",
+            static_cast<uint8_t>(_current_radio_mode),
+            static_cast<uint8_t>(mode));
 
   if (mode == _current_radio_mode && !restart) {
     esp3d_log("Current mode and new mode are identical so cancel");
